Internal Rework.
Before `UTF8_Next()` I made two checks. Now only make one.

In `UTF8_to_Unicode()` and `Unicode_to_UTF8()` I have separated the check and conversion into separate functions. But the functioning of these functions is the same as before.

`UTF8_CharacterSize()` and `Unicode_CharacterSize()` are the functions that are responsible for checking the size of the character.

`UTF8_Converter()` and `Unicode_Converter()` are the functions responsible for the conversion, they require as argument the size of the character to be converted.

Added the variable `size` inside the structure` UTF8_Iterator`. Its use can be both internal and external.